# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Iskele API
  version: 0.1.0-M4
  description: |
    HTTP API of the Iskele daemon (`iskeled`).

    This document is kept in sync with the implementation milestone by
    milestone; only the endpoints listed here exist today. The full planned
    surface is in `PLAN.md` §6.

    **Authentication.** Every endpoint requires a `Bearer` credential — either
    a JWT access token from `/auth/login`, or an API token (`isk_<prefix>_<secret>`)
    — except `/health`, `/version`, `/auth/status`, `/auth/bootstrap`,
    `/auth/login` and `/auth/refresh`.

    Until the first admin account exists, every route except the four auth
    endpoints answers `409 NOT_INITIALIZED`.

    **CSRF.** State-changing requests must carry the `Authorization` header,
    which a cross-site form cannot set, and a present `Origin` must match the
    server. There is no cookie-based session.

    **Streaming.** The WebSocket and SSE endpoints are the exception to the
    Bearer rule: a browser cannot set a header on a WebSocket handshake or an
    `EventSource` request, so they authenticate with a single-use 60-second
    ticket from `POST /auth/ws-ticket`, passed as the `ticket` query parameter.
    They enforce the same permissions as the rest of the API.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://127.0.0.1:8377/api/v1
    description: Default local installation
security:
  - bearerAuth: []
...
paths:
...
  /auth/bootstrap:
    post:
      tags: [auth]
      summary: Create the first admin account
      security: []
      description: |
        Works exactly once, while no account exists. Afterwards it answers
        `409 ALREADY_INITIALIZED` so a passer-by cannot claim the installation.
        Strictly rate limited.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Credentials" }
      responses:
        "201":
          description: The admin account was created and signed in
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          description: An account already exists
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "429":
          $ref: "#/components/responses/RateLimited"
  /auth/login:
    post:
      tags: [auth]
      summary: Sign in
      security: []
      description: |
        Strictly rate limited, and backed by a per-IP lockout after repeated
        failures. A wrong password and an unknown username produce the same
        response, and take the same time, so the endpoint cannot be used to
        enumerate accounts.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Credentials" }
      responses:
        "200":
          description: Signed in
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          description: Invalid username or password
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example:
                error:
                  code: INVALID_CREDENTIALS
                  message: invalid username or password
        "403":
          description: The account is disabled
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409":
          description: The installation has not been set up yet
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "429":
          $ref: "#/components/responses/RateLimited"
  /auth/refresh:
    post:
      tags: [auth]
      summary: Exchange a refresh token for a new pair
      security: []
      description: |
        The refresh token is rotated: the presented one is revoked before the
        new pair is issued, so a stolen copy stops working as soon as the real
        user refreshes.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RefreshRequest" }
      responses:
        "200":
          description: A new token pair
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session" }
        "401":
          description: The refresh token is unknown, expired or already used
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403":
          description: The account is disabled
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /auth/logout:
    post:
      tags: [auth]
      summary: Revoke the current session
      description: |
        Revoking an unknown token is not an error: the caller wanted to be
        signed out, and they are.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RefreshRequest" }
      responses:
        "204":
          description: The session was revoked
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/me:
    get:
      tags: [auth]
      summary: The current identity and its permissions
      description: |
        `permissions` is what the caller's role may do; the UI uses it to hide
        controls rather than showing actions that will be refused.
      responses:
        "200":
          description: The authenticated caller
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CurrentUser" }
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /system/info:
    get:
      tags: [system]
      summary: Docker engine and host summary
      responses:
        "200":
          description: Engine information
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemInfo"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/NotInitialized"
        "503":
          $ref: "#/components/responses/DockerUnavailable"
  /system/df:
    get:
      tags: [system]
      summary: Disk usage (`docker system df`)
      responses:
        "200":
          description: Disk usage report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiskUsage"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/NotInitialized"
        "503":
          $ref: "#/components/responses/DockerUnavailable"
  /containers:
    get:
      tags: [containers]
      summary: List containers
      parameters:
        - name: all
          in: query
          description: Include stopped containers. A bare `?all` counts as true.
          schema: { type: boolean, default: false }
        - name: size
          in: query
          description: Compute `size_rw` / `size_root_fs`. Expensive on large hosts.
          schema: { type: boolean, default: false }
        - name: label
          in: query
          description: Repeatable, or comma-separated. Exact `key=value` match.
          schema: { type: array, items: { type: string } }
        - name: status
          in: query
          description: Repeatable. Engine container states (running, exited, ...).
          schema: { type: array, items: { type: string } }
        - name: name
          in: query
          description: Substring match on the container name.
          schema: { type: string }
      responses:
        "200":
          description: Matching containers
          content:
            application/json:
              schema:
                type: object
                required: [items, total]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Container" }
                  total: { type: integer }
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/NotInitialized"
        "503":
          $ref: "#/components/responses/DockerUnavailable"
  /containers/{id}:
    parameters:
      - $ref: "#/components/parameters/ContainerID"
    get:
      tags: [containers]
      summary: Inspect a container (projected)
      responses:
        "200":
          description: Container detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ContainerDetail" }
        "404":
          $ref: "#/components/responses/ContainerNotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/NotInitialized"
        "503":
          $ref: "#/components/responses/DockerUnavailable"
    delete:
      tags: [containers]
      summary: Remove a container
      parameters:
        - name: force
          in: query
          description: Remove even while running.
          schema: { type: boolean, default: false }
        - name: volumes
          in: query
          description: Also remove anonymous volumes attached to the container.
          schema: { type: boolean, default: false }
      responses:
        "204":
          description: Removed
        "404":
          $ref: "#/components/responses/ContainerNotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "503":
          $ref: "#/components/responses/DockerUnavailable"
...
  /containers/batch:
    post:
      tags: [containers]
      summary: Apply one lifecycle action to several containers
      description: |
        Every container is attempted, whatever happens to the others: a failure
        halfway through a stop-everything must not leave the rest running
        because the loop gave up. The per-container outcome is in `results`.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BatchRequest" }
      responses:
        "200":
          description: Every container succeeded
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BatchResponse" }
        "207":
          description: |
            At least one container failed. The status distinguishes "all done"
            from "mostly done" without the client having to scan every result.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BatchResponse" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/NotInitialized"
        "503":
          $ref: "#/components/responses/DockerUnavailable"
  /auth/ws-ticket:
    post:
      tags: [streaming]
      summary: Issue a single-use ticket for a streaming channel
      description: |
        The ticket carries the caller's identity and permissions, is valid for
        60 seconds, and dies on first use. It is passed as the `ticket` query
        parameter to the WebSocket and SSE endpoints, which is the only place a
        browser can put a credential on those connections.
      responses:
        "201":
          description: Ticket issued
          content:
            application/json:
              schema:
                type: object
                required: [ticket, expires_in]
                properties:
                  ticket: { type: string }
                  expires_in:
                    type: integer
                    description: Seconds until the ticket expires.
                    example: 60
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/NotInitialized"
  /containers/{id}/logs:
    parameters:
      - $ref: "#/components/parameters/ContainerID"
      - $ref: "#/components/parameters/StreamTicket"
      - name: follow
        in: query
        description: |
          Keep the socket open and send new lines as they arrive. Following is
          the point of a WebSocket log stream, so it defaults to true;
          `follow=false` asks for the backlog only.
        schema: { type: boolean, default: true }
      - name: tail
        in: query
        description: |
          Lines of history to replay before following. Defaults to 500:
          replaying a multi-gigabyte log would hang the browser.
        schema: { type: integer, minimum: 0, default: 500 }
      - name: timestamps
        in: query
        description: Include the engine's per-line timestamp.
        schema: { type: boolean, default: false }
      - name: stdout
        in: query
        description: Include stdout. Selecting neither stream means both.
        schema: { type: boolean }
      - name: stderr
        in: query
        description: Include stderr. Selecting neither stream means both.
        schema: { type: boolean }
    get:
      tags: [streaming]
      summary: Stream container logs (WebSocket)
      security: []
      description: |
        A WebSocket endpoint: send `Upgrade: websocket` and authenticate with a
        `ticket`. The server writes JSON text frames shaped like `LogFrame` and
        reads nothing; a client closing the socket stops the engine stream.

        The handshake is rejected when `Origin` does not match `Host`, which is
        the only defense available here — a WebSocket is not subject to the
        same-origin policy the way `fetch` is.

        Frame types: `log` for a line, `err` for a stream failure (the socket
        closes after it), `eof` when the container's stream ends.
      responses:
        "101":
          description: Switching protocols; the log stream follows
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/TicketRejected"
        "403":
          $ref: "#/components/responses/Forbidden"
  /containers/{id}/exec:
    parameters:
      - $ref: "#/components/parameters/ContainerID"
      - $ref: "#/components/parameters/StreamTicket"
      - name: cmd
        in: query
        description: |
          The command to run, one repeat per argument. Defaults to `/bin/sh`,
          the one shell every image has.
        schema: { type: array, items: { type: string } }
      - name: tty
        in: query
        description: Allocate a pseudo-terminal.
        schema: { type: boolean, default: true }
      - name: user
        in: query
        description: Run as this user, e.g. `root` or `1000:1000`.
        schema: { type: string }
      - name: workdir
        in: query
        description: Working directory inside the container.
        schema: { type: string }
      - name: rows
        in: query
        description: Initial terminal height.
        schema: { type: integer, minimum: 1 }
      - name: cols
        in: query
        description: Initial terminal width.
        schema: { type: integer, minimum: 1 }
    get:
      tags: [streaming]
      summary: Interactive exec session (WebSocket)
      security: []
      description: |
        A WebSocket endpoint requiring the `operate` permission. The framing
        carries the channel: **binary** frames are stdin and the command's raw
        output, **text** frames are control messages.

        The only control message is a resize: `{"t":"resize","rows":40,"cols":120}`.
        When the command exits, the server sends a text `{"t":"exit","exit_code":0}`
        frame and closes.

        Every exec session is written to the audit log with the command, the
        user and the container.
      responses:
        "101":
          description: Switching protocols; the exec session follows
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/TicketRejected"
        "403":
          $ref: "#/components/responses/Forbidden"
  /containers/stats:
    parameters:
      - $ref: "#/components/parameters/StreamTicket"
    get:
      tags: [streaming]
      summary: Stream resource usage for every running container (SSE)
      security: []
      description: |
        One `text/event-stream` covering all running containers, each sample
        tagged with its container ID. The list view needs a figure per row, and
        a browser will not open more than six connections to one origin, so the
        daemon fans the per-container engine streams into this one.

        The set of containers is re-scanned every 10 seconds, so one that
        starts appears within that window and one that stops drops out of it. A
        single container failing does not end the stream — only an unreachable
        daemon does.
      responses:
        "200":
          description: The event stream
          content:
            text/event-stream:
              schema: { type: string }
        "401":
          $ref: "#/components/responses/TicketRejected"
        "403":
          $ref: "#/components/responses/Forbidden"
  /containers/{id}/stats:
    parameters:
      - $ref: "#/components/parameters/ContainerID"
      - $ref: "#/components/parameters/StreamTicket"
    get:
      tags: [streaming]
      summary: Stream resource usage (SSE)
      security: []
      description: |
        A `text/event-stream` of one sample per second. Event names: `stats`
        with a `Stats` payload, `error` with `{code, message}`, and `eof` when
        the container stops. A comment line is sent every 25 seconds so an idle
        stream is not closed by a proxy.

        CPU percentage is computed the way `docker stats` computes it: the
        container's CPU delta over the system delta, multiplied by the number
        of cores. Memory usage excludes the page cache, so the number matches
        what the kernel would reclaim under pressure rather than the raw
        cgroup total.
      responses:
        "200":
          description: The event stream
          content:
            text/event-stream:
              schema: { type: string }
        "401":
          $ref: "#/components/responses/TicketRejected"
        "403":
          $ref: "#/components/responses/Forbidden"
  /system/events:
    parameters:
      - $ref: "#/components/parameters/StreamTicket"
    get:
      tags: [streaming]
      summary: Stream Docker engine events (SSE)
      security: []
      description: |
        A `text/event-stream` of engine events — containers starting and
        dying, images being pulled, volumes and networks changing. The UI uses
        it to refresh lists without polling. Event names: `event` with a
        `DockerEvent` payload, `error` with `{code, message}`.
      responses:
        "200":
          description: The event stream
          content:
            text/event-stream:
              schema: { type: string }
        "401":
          $ref: "#/components/responses/TicketRejected"
        "403":
          $ref: "#/components/responses/Forbidden"
  /images:
    get:
      tags: [images]
      summary: List images
      parameters:
        - name: all
          in: query
          description: Include intermediate layers.
          schema: { type: boolean, default: false }
        - name: dangling
          in: query
          description: |
            Tri-state. Absent means no filter; `true` restricts to untagged
            images, `false` to tagged ones.
          schema: { type: boolean }
        - name: label
          in: query
          schema: { type: array, items: { type: string } }
      responses:
        "200":
          description: Matching images, newest first
          content:
            application/json:
              schema:
                type: object
                required: [items, total]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Image" }
                  total: { type: integer }
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/NotInitialized"
        "503":
          $ref: "#/components/responses/DockerUnavailable"
  /volumes:
    get:
      tags: [volumes]
      summary: List volumes
      responses:
        "200":
          description: Volumes, sorted by name
          content:
            application/json:
              schema:
                type: object
                required: [items, total]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Volume" }
                  total: { type: integer }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/NotInitialized"
        "503":
          $ref: "#/components/responses/DockerUnavailable"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        A JWT access token from `/auth/login`, or an API token of the form
        `isk_<prefix>_<secret>`. API tokens carry their owner's role.
...
  responses:
...
    BadRequest:
      description: A query parameter was malformed
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
...
    Credentials:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          description: Case-insensitive; the casing given is preserved for display.
        password:
          type: string
          format: password
          minLength: 12
          maxLength: 1024
          description: |
            At least 12 characters and at least two character classes
            (lowercase, uppercase, digit, symbol).
...